Reorganize testing and host stuff.
