+++ stellar_contract_sdk/src/testing.rs
@@ -1,0 +1,8 @@
+#![allow(unused_variables)]
+
+pub mod fs;
+pub(crate) mod host;
+pub mod mem;
+
+pub use host::swap_mock_host;
+pub use host::MockHost;
