+++ stellar_contract_sdk/src/testing/host.rs
@@ -4,12 +4,28 @@
 // copy of either the sandbox or the real implementation in stellar-core, but it
 // has semantics that are as close as possible.
 
-#![allow(unused_variables)]
+use crate::{Object, Status, Val};
 
-pub mod fs;
-pub mod mem;
+use core::fmt::Debug;
 
-use crate::{Object, Val};
+impl Debug for Val {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Val")
+            .field("body", &self.get_body())
+            .field("tag", &self.get_tag())
+            .finish()
+    }
+}
+
+impl core::fmt::Debug for Status {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let (ty, code) = self.decompose_status();
+        f.debug_struct("Status")
+            .field("type", &ty)
+            .field("code", &code)
+            .finish()
+    }
+}
 
 pub trait MockHost {
     fn log_value(&mut self, v: Val) -> Val;
@@ -46,7 +62,7 @@ pub trait MockHost {
 use core::cell::RefCell;
 thread_local! {
     // The default mock host is a memory-backed one, but the user can replace it if they like.
-    pub static MOCK_HOST: RefCell<Box<dyn MockHost>> = RefCell::new(Box::new(mem::MemHost::new()));
+    pub static MOCK_HOST: RefCell<Box<dyn MockHost>> = RefCell::new(Box::new(super::mem::MemHost::new()));
 }
 
 // Replaces the current mock host with `mock`, returning the previous mock.
