+++ stellar_contract_sdk/src/testing/mem.rs
