--- stellar_contract_sdk/src/testing/mod.rs
@@ -1,8 +1,0 @@
-#[cfg(not(target_family = "wasm"))]
-pub mod host;
-
-#[cfg(not(target_family = "wasm"))]
-mod val;
-
-#[cfg(not(target_family = "wasm"))]
-mod status;
